# Do not reencode TSDB head chunks when querying ingesters