--- a/pkg/ingester/ingester_v2.go
+++ b/pkg/ingester/ingester_v2.go
@@ -1467,8 +1467,12 @@ func (i *Ingester) v2QueryStreamSamples(ctx context.Context, db *userTSDB, from,
 	}
 	defer q.Close()
 
+	var hints *storage.SelectHints
+	if shard != nil {
+		hints = configSelectHintsWithShard(initSelectHints(from, through), shard)
+	}
+
 	// It's not required to return sorted series because series are sorted by the Mimir querier.
-	hints := getSelectHintsForShard(from, through, shard)
 	ss := q.Select(false, hints, matchers...)
 	if ss.Err() != nil {
 		return 0, 0, ss.Err()
@@ -1537,8 +1541,13 @@ func (i *Ingester) v2QueryStreamChunks(ctx context.Context, db *userTSDB, from, 
 	}
 	defer q.Close()
 
+	// Disable chunks trimming, so that we don't have to rewrite chunks which have samples outside
+	// the requested from/through range. PromQL engine can handle it.
+	hints := initSelectHints(from, through)
+	hints = configSelectHintsWithShard(hints, shard)
+	hints = configSelectHintsWithTrimDisabled(hints)
+
 	// It's not required to return sorted series because series are sorted by the Mimir querier.
-	hints := getSelectHintsForShard(from, through, shard)
 	ss := q.Select(false, hints, matchers...)
 	if ss.Err() != nil {
 		return 0, 0, ss.Err()
@@ -1574,6 +1583,8 @@ func (i *Ingester) v2QueryStreamChunks(ctx context.Context, db *userTSDB, from, 
 			switch meta.Chunk.Encoding() {
 			case chunkenc.EncXOR:
 				ch.Encoding = int32(encoding.PrometheusXorChunk)
+			case chunkenc.EncXORPartial:
+				ch.Encoding = int32(encoding.PrometheusXorChunkPartial)
 			default:
 				return 0, 0, errors.Errorf("unknown chunk encoding from TSDB chunk querier: %v", meta.Chunk.Encoding())
 			}
@@ -2397,16 +2408,23 @@ func (i *Ingester) getInstanceLimits() *InstanceLimits {
 	return l
 }
 
-func getSelectHintsForShard(start, end int64, shard *querysharding.ShardSelector) *storage.SelectHints {
-	if shard == nil {
-		return nil
-	}
-
-	// If query sharding is enabled, we need to pass it along with hints.
+func initSelectHints(start, end int64) *storage.SelectHints {
 	return &storage.SelectHints{
-		Start:      start,
-		End:        end,
-		ShardIndex: shard.ShardIndex,
-		ShardCount: shard.ShardCount,
+		Start: start,
+		End:   end,
 	}
+}
+
+func configSelectHintsWithShard(hints *storage.SelectHints, shard *querysharding.ShardSelector) *storage.SelectHints {
+	if shard != nil {
+		// If query sharding is enabled, we need to pass it along with hints.
+		hints.ShardIndex = shard.ShardIndex
+		hints.ShardCount = shard.ShardCount
+	}
+	return hints
+}
+
+func configSelectHintsWithTrimDisabled(hints *storage.SelectHints) *storage.SelectHints {
+	hints.TrimDisabled = true
+	return hints
 }