**What this PR does**:
This PR builds on top of https://github.com/grafana/mimir/pull/430.

When querying chunks from TSDB head, the chunks which are still open (the ones we're still appending to and not mmap-ed yet) are re-encoded for every single `querier.Select()` call. This puts a significative pressure on ingesters on high query load, both in terms of CPU and memory.

For example, from this profile we can count almost 20% of memory allocations (bytes) in the profiled ingester due to this:

![Screenshot 2021-10-28 at 17 58 42](https://user-images.githubusercontent.com/1701904/139292681-73f54d3c-daa4-4f4e-8c91-44238621ae48.png)

In this PR I'm showing an experiment to avoid having to reencode the TSDB head chunks. TSDB is quite tricky and I'm not sure I'm missing anything (a part from the fact I intentionally didn't support isolation yet, to keep the experiment simpler). TSDB changes are [here](https://github.com/grafana/prometheus-private/compare/experiment-not-reencode-partial-chunks) (on top of changes done for #430 too).

**Benchmark** (including optimization proposed in #430 which is required to have this PR optimization working):

We don't expect any improvement for querying samples:

```
name                                                                                           old time/op    new time/op    delta
Ingester_V2QueryStream/query_samples/time_range=full_data,_query_sharding=false-12                868ms ± 2%     803ms ± 2%   -7.43%  (p=0.003 n=3+3)
Ingester_V2QueryStream/query_samples/time_range=full_data,_query_sharding=true-12                 921ms ± 0%     857ms ± 1%   -6.96%  (p=0.006 n=3+3)
Ingester_V2QueryStream/query_samples/time_range=partial_block,_query_sharding=false-12            174ms ± 0%     171ms ± 3%     ~     (p=0.254 n=3+3)
Ingester_V2QueryStream/query_samples/time_range=partial_block,_query_sharding=true-12             207ms ± 4%     197ms ± 1%     ~     (p=0.151 n=3+3)
Ingester_V2QueryStream/query_samples/time_range=partial_head,_query_sharding=false-12             396ms ± 5%     382ms ± 1%     ~     (p=0.263 n=3+3)
Ingester_V2QueryStream/query_samples/time_range=partial_head,_query_sharding=true-12              427ms ± 2%     413ms ± 3%     ~     (p=0.112 n=3+3)
Ingester_V2QueryStream/query_samples/time_range=head_+_partial_block,_query_sharding=false-12     828ms ± 2%     809ms ± 2%     ~     (p=0.149 n=3+3)
Ingester_V2QueryStream/query_samples/time_range=head_+_partial_block,_query_sharding=true-12      907ms ± 3%     861ms ± 2%     ~     (p=0.084 n=3+3)
Ingester_V2QueryStream/query_chunks/time_range=full_data,_query_sharding=false-12                 370ms ± 5%     154ms ± 5%  -58.53%  (p=0.001 n=3+3)
Ingester_V2QueryStream/query_chunks/time_range=full_data,_query_sharding=true-12                  414ms ± 2%     205ms ± 2%  -50.55%  (p=0.000 n=3+3)
Ingester_V2QueryStream/query_chunks/time_range=partial_block,_query_sharding=false-12             193ms ± 2%      37ms ± 2%  -80.79%  (p=0.000 n=3+3)
Ingester_V2QueryStream/query_chunks/time_range=partial_block,_query_sharding=true-12              216ms ± 4%      60ms ± 5%  -72.34%  (p=0.000 n=3+3)
Ingester_V2QueryStream/query_chunks/time_range=partial_head,_query_sharding=false-12              471ms ± 2%      49ms ± 3%  -89.67%  (p=0.000 n=3+3)
Ingester_V2QueryStream/query_chunks/time_range=partial_head,_query_sharding=true-12               495ms ± 5%      79ms ± 5%  -84.11%  (p=0.001 n=3+3)
Ingester_V2QueryStream/query_chunks/time_range=head_+_partial_block,_query_sharding=false-12      501ms ± 7%     141ms ± 1%  -71.81%  (p=0.003 n=3+3)
Ingester_V2QueryStream/query_chunks/time_range=head_+_partial_block,_query_sharding=true-12       585ms ±15%     204ms ± 1%  -65.09%  (p=0.014 n=3+3)

name                                                                                           old alloc/op   new alloc/op   delta
Ingester_V2QueryStream/query_samples/time_range=full_data,_query_sharding=false-12                478MB ± 0%     480MB ± 0%   +0.59%  (p=0.000 n=3+3)
Ingester_V2QueryStream/query_samples/time_range=full_data,_query_sharding=true-12                 479MB ± 0%     481MB ± 0%   +0.58%  (p=0.000 n=3+3)
Ingester_V2QueryStream/query_samples/time_range=partial_block,_query_sharding=false-12            121MB ± 0%     121MB ± 0%     ~     (p=0.640 n=3+3)
Ingester_V2QueryStream/query_samples/time_range=partial_block,_query_sharding=true-12             122MB ± 0%     122MB ± 0%     ~     (p=0.401 n=3+3)
Ingester_V2QueryStream/query_samples/time_range=partial_head,_query_sharding=false-12             239MB ± 0%     239MB ± 0%     ~     (p=0.663 n=3+3)
Ingester_V2QueryStream/query_samples/time_range=partial_head,_query_sharding=true-12              240MB ± 0%     240MB ± 0%     ~     (p=0.202 n=3+3)
Ingester_V2QueryStream/query_samples/time_range=head_+_partial_block,_query_sharding=false-12     479MB ± 0%     479MB ± 0%     ~     (p=0.178 n=3+3)
Ingester_V2QueryStream/query_samples/time_range=head_+_partial_block,_query_sharding=true-12      480MB ± 0%     480MB ± 0%     ~     (p=0.464 n=3+3)
Ingester_V2QueryStream/query_chunks/time_range=full_data,_query_sharding=false-12                81.2MB ± 0%    69.6MB ± 0%  -14.28%  (p=0.000 n=3+3)
Ingester_V2QueryStream/query_chunks/time_range=full_data,_query_sharding=true-12                 82.2MB ± 0%    70.6MB ± 0%  -14.11%  (p=0.000 n=3+3)
Ingester_V2QueryStream/query_chunks/time_range=partial_block,_query_sharding=false-12            28.8MB ± 0%    15.6MB ± 0%  -45.80%  (p=0.000 n=3+3)
Ingester_V2QueryStream/query_chunks/time_range=partial_block,_query_sharding=true-12             29.7MB ± 0%    16.5MB ± 0%  -44.51%  (p=0.000 n=3+3)
Ingester_V2QueryStream/query_chunks/time_range=partial_head,_query_sharding=false-12             67.8MB ± 0%    32.1MB ± 0%  -52.64%  (p=0.000 n=3+3)
Ingester_V2QueryStream/query_chunks/time_range=partial_head,_query_sharding=true-12              68.6MB ± 0%    32.9MB ± 0%  -52.01%  (p=0.000 n=3+3)
Ingester_V2QueryStream/query_chunks/time_range=head_+_partial_block,_query_sharding=false-12     93.6MB ± 0%    69.6MB ± 0%  -25.64%  (p=0.000 n=3+3)
Ingester_V2QueryStream/query_chunks/time_range=head_+_partial_block,_query_sharding=true-12      94.6MB ± 0%    70.6MB ± 0%  -25.37%  (p=0.000 n=3+3)

name                                                                                           old allocs/op  new allocs/op  delta
Ingester_V2QueryStream/query_samples/time_range=full_data,_query_sharding=false-12                1.68M ± 0%     1.73M ± 0%   +2.97%  (p=0.000 n=3+3)
Ingester_V2QueryStream/query_samples/time_range=full_data,_query_sharding=true-12                 1.69M ± 0%     1.74M ± 0%   +2.96%  (p=0.000 n=3+3)
Ingester_V2QueryStream/query_samples/time_range=partial_block,_query_sharding=false-12             650k ± 0%      650k ± 0%     ~     (p=0.577 n=3+3)
Ingester_V2QueryStream/query_samples/time_range=partial_block,_query_sharding=true-12              651k ± 0%      651k ± 0%     ~     (p=0.457 n=3+3)
Ingester_V2QueryStream/query_samples/time_range=partial_head,_query_sharding=false-12              834k ± 0%      834k ± 0%     ~     (p=0.591 n=3+3)
Ingester_V2QueryStream/query_samples/time_range=partial_head,_query_sharding=true-12               835k ± 0%      835k ± 0%     ~     (p=0.267 n=3+3)
Ingester_V2QueryStream/query_samples/time_range=head_+_partial_block,_query_sharding=false-12     1.78M ± 0%     1.78M ± 0%     ~     (p=0.132 n=3+3)
Ingester_V2QueryStream/query_samples/time_range=head_+_partial_block,_query_sharding=true-12      1.79M ± 0%     1.79M ± 0%     ~     (p=0.496 n=3+3)
Ingester_V2QueryStream/query_chunks/time_range=full_data,_query_sharding=false-12                 1.49M ± 0%     1.37M ± 0%   -8.38%  (p=0.000 n=3+3)
Ingester_V2QueryStream/query_chunks/time_range=full_data,_query_sharding=true-12                  1.49M ± 0%     1.37M ± 0%   -8.37%  (p=0.000 n=3+3)
Ingester_V2QueryStream/query_chunks/time_range=partial_block,_query_sharding=false-12              575k ± 0%      350k ± 0%  -39.11%  (p=0.000 n=3+3)
Ingester_V2QueryStream/query_chunks/time_range=partial_block,_query_sharding=true-12               576k ± 0%      351k ± 0%  -39.04%  (p=0.000 n=3+3)
Ingester_V2QueryStream/query_chunks/time_range=partial_head,_query_sharding=false-12               909k ± 0%      442k ± 0%  -51.36%  (p=0.000 n=3+3)
Ingester_V2QueryStream/query_chunks/time_range=partial_head,_query_sharding=true-12                910k ± 0%      443k ± 0%  -51.30%  (p=0.000 n=3+3)
Ingester_V2QueryStream/query_chunks/time_range=head_+_partial_block,_query_sharding=false-12      1.72M ± 0%     1.37M ± 0%  -20.38%  (p=0.000 n=3+3)
Ingester_V2QueryStream/query_chunks/time_range=head_+_partial_block,_query_sharding=true-12       1.72M ± 0%     1.37M ± 0%  -20.36%  (p=0.000 n=3+3)
```

**Draft** because:
- Very experimental, just to prove the idea
- Missing a bunch of tests (correctness, race, ...)

**Which issue(s) this PR fixes**:
N/A

**Checklist**

- [ ] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`